+++ pkg/server/conn_stmt_test.go
@@ -200,7 +200,7 @@ func TestCursorWithParams(t *testing.T) {
 
 		require.Equal(t, "rg1", reporter.tidbGroup)
 		expectedCursorDelta := ruv2Metrics.CalculateRUValues(weights) - baselineTiDBRU
-		require.Equal(t, float64(expectedCursorDelta), reporter.tidbRUV2)
+		require.Equal(t, expectedCursorDelta, reporter.tidbRUV2)
 		require.Equal(t, 0.0, reporter.tikvRUV2)
 
 		ruDetails.AddTiKVRUV2(7)
